+++ src/main/java/br/com/zalf/prolog/webservice/geral/unidade/_model/UnidadeEdicao.java
@@ -1,5 +1,7 @@
 package br.com.zalf.prolog.webservice.geral.unidade._model;
 
+import io.swagger.annotations.ApiModel;
+import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 import org.jetbrains.annotations.Nullable;
 
@@ -11,8 +13,16 @@ import javax.validation.constraints.Size;
  *
  * @author Gustavo Navarro (https://github.com/gustavocnp95)
  */
+@ApiModel(value = "Unidade", description = "Dados a serem atualizados em uma unidade")
 @Data
 public final class UnidadeEdicao {
+
+    @ApiModelProperty(
+            value = "Um código de unidade, obrigatório",
+            name = "codUnidade",
+            dataType = "Long",
+            example = "215"
+    )
     @NotNull(message = "O código da unidade é obrigatório.")
     private final Long codUnidade;
     @NotNull(message = "O nome da unidade não pode estar vazio.")